Move File class into it's own file and move util.js & ingester.js into a src folder
