+++ test/ingestor.tests.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect;
-const { filewalker, state } = require('../ingestor.js');
+const { filewalker, state } = require('../src/ingestor.js');
 
 describe('File Walker Ingestor', () => {
   const testDirectory = `${__dirname}/testArchive`;
@@ -16,7 +16,7 @@ describe('File Walker Ingestor', () => {
     });
   });
 
-  it('it should ouput a messages array with a length of 9', () => {
+  it('it should ouput a messages array with a length of 3', () => {
     expect(testState.messages).to.have.lengthOf(3);
     expect(testState.messages[0].text).to.be.equal('testMessage1');
   });
