+++ test/utils.tests.js
@@ -1,4 +1,5 @@
-const { File, processFileController } = require('../utils.js');
+const { processFileController } = require('../src/utils.js');
+const File = require('../src/File').File;
 const chai = require('chai');
 const chaiAsPromised = require('chai-as-promised');
 
@@ -12,12 +13,12 @@ describe('File Class', () => {
   const newFile = new File(testFilePath);
 
   it('it should be an object that containts a filename property', () => {
-    expect(newFile).to.be.an('object');
     expect(newFile.filename).to.be.equal(testFilePath);
+    //Testing a returned promise
   });
 
   it('it should read a file and parse the json data and store in the filecontents property', async () => {
-    const results = await newFile.readFile();
+    const results = await newFile.getFileContents();
     expect(results)
       .to.be.an('object')
       .to.have.property('test');
@@ -26,7 +27,7 @@ describe('File Class', () => {
 
   it('it should catch an error if the file path is incorrect', async () => {
     const newFile = new File(invalidFilePath);
-    await expect(newFile.readFile()).to.be.rejected;
+    await expect(newFile.getFileContents()).to.be.rejected;
   });
 });
 describe('fileProcessController Utility', () => {
